// components/shared/Metric/Metric.interface.ts
export interface MetricProps {
	imgUrl: string; // URL зображення
	alt: string; // Альтернативний текст для зображення
	value: string | number; // Значення, яке буде відображено в абзаці.
	title: string; // Заголовок, який буде відображено в <span> всередині абзацу
	href?: string; // Опціональний URL для посилання, якщо задано, додається додатковий клас до зображення
	textStyles?: string; // Опціональні додаткові класи для стилізації абзацу
	isAuthor?: boolean; // Опціональний прапорець, якщо true, додається додатковий клас до <span>
}